added random ids, made formatJSON() a class function

+++ api/api.py
@@ -2,17 +2,20 @@ from time import time
 from flask import Flask, request
 from datetime import datetime
 from flask_sqlalchemy import SQLAlchemy
+import random, string
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://postgres:hellothere@localhost/when-works'
 db = SQLAlchemy(app)
 
+def gen_key():
+	return ''.join(random.choices(string.ascii_letters + string.digits, k=7))
+
 class Event(db.Model):
-	id = db.Column(db.Integer, primary_key=True)
+	id = db.Column(db.String(7), default=gen_key(), primary_key=True)
 	event_name = db.Column(db.String(100), nullable=False)
 	host_name = db.Column(db.String(100), nullable=False)
 	created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-	num_guests = db.Column(db.Integer, default=1, nullable=False)
 	guests = db.relationship('Guest', backref='event')
 
 
@@ -23,15 +26,32 @@ class Event(db.Model):
 		self.event_name = event_name
 		self.host_name = host_name
 
+	def formatJSON(self):
+		names = []
+		guest_ids = []
+		for guest in self.guests:
+			names.append(guest.name)
+			guest_ids.append(guest.id)
+
+		return {
+			"id": self.id,
+			"event_name": self.event_name,
+			"host_name": self.host_name,
+			"created_at": self.created_at,
+			"guests": names,
+			"guest_ids": guest_ids
+		}
+
+
 guest_time = db.Table('guest_time', 
-	db.Column('guest_id', db.Integer, db.ForeignKey('guest.id')),
+	db.Column('guest_id', db.String(7), db.ForeignKey('guest.id')),
 	db.Column('time_id', db.DateTime, db.ForeignKey('time.time_id'))
 )
 
 class Guest(db.Model):
-	id = db.Column(db.Integer, primary_key=True)
+	id = db.Column(db.String(7), default=gen_key(), primary_key=True)
 	name = db.Column(db.String(100), nullable=False)
-	event_id = db.Column(db.Integer, db.ForeignKey('event.id'), nullable=False)
+	event_id = db.Column(db.String(7), db.ForeignKey('event.id'), nullable=False)
 	available_times = db.relationship('Time', secondary=guest_time, backref = 'guests')
 
 	def __repr__(self):
@@ -41,6 +61,20 @@ class Guest(db.Model):
 		self.name = name
 		self.event_id = event_id
 
+	def formatJSON(self):
+		times = []
+		for time in self.available_times:
+			times.append(str(time))
+
+		return {
+			"id": self.id,
+			"guest_name": self.name,
+			"event_id": self.event_id,
+			"event_name": (Event.query.filter_by(id=self.event_id).first()).event_name,
+			"avail_times": times
+	}
+
+
 class Time(db.Model):
 	time_id = db.Column(db.DateTime, primary_key=True, nullable = False)
 
@@ -49,36 +83,6 @@ class Time(db.Model):
 
 	def __init__(self, time):
 		self.time_id = time
-
-def format_event(event):
-	names = []
-	guests = Guest.query.filter_by(event_id=event.id)
-	for guest in guests:
-		names.append(guest.name)
-
-	return {
-		"id": event.id,
-		"event_name": event.event_name,
-		"host_name": event.host_name,
-		"num_guests": event.num_guests,
-		"created_at": event.created_at,
-		"guests": names
-	}
-
-
-def format_guest(guest):
-	times = []
-	avail_times = guest.available_times
-	for time in avail_times:
-		times.append(str(time))
-
-	return {
-		"id": guest.id,
-		"guest_name": guest.name,
-		"event_id": guest.event_id,
-		"event_name": (Event.query.filter_by(id=guest.event_id).first()).event_name,
-		"avail_times": times
-	}
 
 @app.route('/time')
 def get_current_time():
@@ -103,19 +107,20 @@ def create_event():
 		if (tmp_time is None):
 			db.session.add(Time(time))
 			tmp_time = Time.query.filter_by(time_id=time).first()
+
+		# add relationship to join table
 		guest.available_times.append(tmp_time)
 
 	db.session.add(guest)
-	event.num_guests += 1
 	db.session.commit()
-	return format_event(event)
+	return event.formatJSON()
 
 @app.route('/event', methods=['GET'])
 def get_events():
 	events = Event.query.order_by(Event.id.asc()).all()
 	event_list = []
 	for event in events:
-		event_list.append(format_event(event))
+		event_list.append(event.formatJSON())
 	return {'events': event_list}
 
 @app.route('/<event_id>', methods=['POST'])
@@ -129,50 +134,73 @@ def create_guest(event_id):
 	# add guest to database
 	guest = Guest(name, event_id)
 
-	# add time to database if it doesnt exist, and add
-	#	relationship to join table
+	# add time to database
 	for time in avail_times:
+		# query if this time exists in the time database
 		tmp_time = Time.query.filter_by(time_id=time).first()
 		if (tmp_time is None):
 			db.session.add(Time(time))
 			tmp_time = Time.query.filter_by(time_id=time).first()
+
+		# add relationship to join table
 		guest.available_times.append(tmp_time)
 
 	db.session.add(guest)
-	# event is initialized with num_guests = 1
 	db.session.commit()
-	return format_guest(guest)
+	return guest.formatJSON()
 
 @app.route('/<event_id>', methods=['GET'])
 def get_event(event_id):
 	event = Event.query.filter_by(id=event_id).first()
 	if (event is None): return f"Event ID {event_id} does not exist"
-	return format_event(event)
+	return event.formatJSON()
+
+@app.route('/<event_id>/<guest_id>', methods=['GET'])
+def get_guest(event_id, guest_id):
+	guest = Guest.query.filter_by(id=guest_id).first()
+	if (guest == None or guest.event_id != event_id):
+		return "This guest either does not exist, or does not belong to this event.\n"
+
+	return guest.formatJSON()
+
+@app.route('/<event_id>/guests', methods=['GET'])
+def get_guests(event_id):
+	event = Event.query.filter_by(id=event_id).first()
+	if (event == None): return "Event does not exist."
+	
+	guest_list = []
+	for guest in event.guests:
+		guest_list.append(guest.formatJSON())
+
+	return guest_list
 
 @app.route('/<event_id>', methods=['DELETE'])
 def delete_event(event_id):
 	event = Event.query.filter_by(id=event_id).first()
 	if (event is None): return "Event does not exist"
 	guests = Guest.query.filter_by(event_id=event_id)
+
 	for guest in guests:
 		db.session.delete(guest)
+
 	db.session.delete(event)
 	db.session.commit()
+
 	return 'Event Removed'
 
 @app.route('/<event_id>/<guest_id>', methods=['DELETE'])
 def delete_guest(event_id, guest_id):
 	event = Event.query.filter_by(id=event_id).first()
+	if (event is None): return "Event does not exist"
+
 	guest = Guest.query.filter_by(id=guest_id).first()
-	event.num_guests -= 1
+	if (guest is None):
+		return 'Guest does not exist'
+
+	if (guest.event_id != event_id):
+		return 'Guest does not belong to this event'
+
 	db.session.delete(guest)
 	db.session.commit()
 	return 'Guest Removed'
-
-@app.route('/<event_id>/<guest_id>', methods=['GET'])
-def get_guest(event_id, guest_id):
-	guest = Guest.query.filter_by(id=guest_id).first()
-	if (guest == None or guest.event_id is not event_id):
-		return "This guest either does not exist, or does not belong to this event.\n"
-	return format_guest(guest)
